Remove player classes
Hangman will only have one class now the "hangman class". All functions of the CPU class have been moved to the hangman class.

Furthermore, this commit when implemented adds the #get_letter_or_guess function. This function will get user input form the player. User input must be a single letter or a word that is the same length as the secret word.